Updated statistics

+++ classes.py
@@ -27,7 +27,7 @@ def change_exits(tiles, coord, size, orientation):
                 borders.add(tile)
                 tile.image = tile_images['floor']
 
-    exits.remove(*exits)
+    exits.empty()
 
 
 def change_doors(tiles):
@@ -40,15 +40,16 @@ def change_enemies(tiles):
     current_enemies.add(*tiles)
 
 
-ENEMY_DELAY = 500
+ENEMY_DELAY = 1000
 ENEMY_EVENT_TYPE = pygame.USEREVENT + 1
 MAX_HEALTH = 10
 MAX_ARMOR = 5
 MAX_MANA = 100
-HERO_GET_ARMOR = pygame.USEREVENT + 1
-HERO_GET_MANA = pygame.USEREVENT + 1
+HERO_GET_ARMOR = ENEMY_EVENT_TYPE + 1
+HERO_GET_MANA = HERO_GET_ARMOR + 1
 ARMOR_DELAY = 20000
 MANA_DELAY = 10000
+OPEN_DOORS_EVENT = pygame.event.Event(HERO_GET_MANA + 1)
 entries = pygame.sprite.Group()
 current_entries = pygame.sprite.Group()
 doors = pygame.sprite.Group()
@@ -133,9 +134,13 @@ class Room(pygame.sprite.Group):
                 elif self.map[j][i] == 'w':
                     Wall('classwall', x, y, self, borders)
 
-                elif self.map[j][i] == '+':
+                elif self.map[j][i] == '~':
                     Tile('parquet', x, y, self)
-                    enemies.add(BaseEnemy((x, y), 10, 1))
+                    enemies.add(CloseEnemy((x, y), 5, 1, 2))
+
+                elif self.map[j][i] == '^':
+                    Tile('parquet', x, y, self)
+                    enemies.add(DistanceEnemy((x, y), 5, 1))
 
                 elif self.map[j][i] == '0':
                     Tile('empty', x, y, self)
@@ -158,12 +163,12 @@ class Room(pygame.sprite.Group):
 
 class Map:
     def __init__(self):
-        start_height, start_width = 11, 16
+        start_height, start_width = 15, 16
         corridor_width = 9
         x, y = (WIDTH - (start_width - 1) * TILE_WIDTH) // 2,\
                (HEIGHT - (start_height - 1) * TILE_HEIGHT) // 2
         corridor_pos = x + start_width * TILE_WIDTH, y + ((start_height // 2) - 2) * TILE_HEIGHT
-        classroom_pos = corridor_pos[0] + TILE_WIDTH * corridor_width, y - 3 * TILE_HEIGHT
+        classroom_pos = corridor_pos[0] + TILE_WIDTH * corridor_width, y - TILE_HEIGHT
 
         self.map = [Room('start.txt', (x, y)),
                     Room('horizontal.txt', corridor_pos),
@@ -193,6 +198,7 @@ class Map:
 
         else:  # новая комната и коридор к ней ниже
             change_exits(exits, room.y, room.height * TILE_HEIGHT, 'v')
+            change_enemies(enemies)
             x, y = room.x, room.y
             width, height = room.width, room.height
             corridor_pos = x + (width // 2 - 2) * TILE_WIDTH, y + height * TILE_HEIGHT
@@ -209,7 +215,7 @@ class BaseEntity(pygame.sprite.Sprite):
         self.radius = TILE_WIDTH * 7 // 10
         x, y = position
         # Rect для взаимодействия объекта со стенами
-        self.rect = pygame.Rect(x, y + self.radius // 2, self.radius, self.radius // 2)
+        self.rect = pygame.Rect(x, y + self.radius // 4, self.radius, self.radius // 2)
 
     def get_position(self):
         return self.rect.center
@@ -229,15 +235,20 @@ class Hero(BaseEntity):
 
     def render(self, screen):
         pygame.draw.rect(screen, (255, 255, 255), self.body_rect)
+        pygame.draw.rect(screen, (0, 0, 0), self.rect)
 
     def attack(self):
-        x1, y1, = self.rect.center
-        x2, y2 = pygame.mouse.get_pos()
-        vector = (x2 - x1, y2 - y1)
-        length = math.sqrt(vector[0] ** 2 + vector[1] ** 2)
-        shells.add(Shell(self.rect.center, vector, length, True, 2))
-        if length != 0:
-            self.mana -= 1
+        if self.mana:
+            x1, y1, = self.rect.center
+            x2, y2 = pygame.mouse.get_pos()
+            vector = (x2 - x1, y2 - y1)
+            length = math.sqrt(vector[0] ** 2 + vector[1] ** 2)
+            shells.add(Shell(self.rect.center, vector, length, True, 2))
+            if length != 0:
+                self.mana -= 1
+
+    def delete(self):
+        pass
 
 
 class BaseEnemy(BaseEntity):
@@ -246,40 +257,11 @@ class BaseEnemy(BaseEntity):
         x, y = position
         self.health = health
         self.power = power
-        self.speed = 4
-        self.dx, self.dy = 0, 0
         self.delay = 500
-        self.steps = 0
         self.is_alive = True
 
         # Rect для отрисовки и улавливания снарядов противников
         self.body_rect = pygame.Rect(x, y - self.radius, self.radius, 2 * self.radius)
-
-    def move(self):
-        if self.steps and self.is_alive:
-            x, y = self.get_position()
-            x += self.dx
-            y += self.dy
-            if not pygame.sprite.spritecollideany(BaseEntity((x, y)), borders):
-                self.steps -= 1
-                self.rect.center = x, y
-                self.body_rect.center = x, y
-
-    def go_to(self, position):
-        if self.is_alive:
-            x1, y1, = self.rect.center
-            x2, y2 = position
-            vector = (x2 - x1, y2 - y1)
-            length = math.sqrt(vector[0] ** 2 + vector[1] ** 2)
-            if length != 0:
-                self.dx, self.dy, self.steps = \
-                    self.speed * vector[0] / length,\
-                    self.speed * vector[1] / length, length // self.speed
-            else:
-                self.steps = 0
-
-    def attack(self, position):
-        pass
 
     def render(self, screen):
         pygame.draw.rect(screen, (100, 100, 100), self.body_rect)
@@ -293,8 +275,64 @@ class BaseEnemy(BaseEntity):
         global total_kills
 
         self.is_alive = False
+        self.body_rect.h = self.body_rect.h // 2
+        if all(map(lambda enemy: not enemy.is_alive, current_enemies.sprites())):
+            pygame.event.post(OPEN_DOORS_EVENT)
         total_kills += 1
-        self.body_rect.y += self.body_rect.h // 2
+
+
+class CloseEnemy(BaseEnemy):
+    def __init__(self, position, health, power, speed):
+        super().__init__(position, health, power)
+        self.speed = speed
+        self.dx, self.dy = 0, 0
+        self.steps = 0
+
+    def move(self):
+        if self.steps and self.is_alive:
+            x, y = self.rect.topleft
+            x += self.dx
+            y += self.dy
+            if not pygame.sprite.spritecollideany(BaseEntity((x, y)), borders):
+                self.steps -= 1
+                self.rect.topleft = x, y
+                self.body_rect.topleft = x, y - self.radius
+
+    def action(self, hero):
+        if self.is_alive:
+            x1, y1, = self.rect.center
+            x2, y2 = hero.get_position()
+            vector = (x2 - x1, y2 - y1)
+            length = math.sqrt(vector[0] ** 2 + vector[1] ** 2)
+            if length > 20:
+                self.dx, self.dy, self.steps = \
+                    self.speed * vector[0] / length, \
+                    self.speed * vector[1] / length, length // self.speed
+            else:
+                self.steps = 0
+                if hero.armor > self.power:
+                    hero.armor -= self.power
+                else:
+                    hero.armor = 0
+                    damage = self.power - hero.armor
+                    if hero.health > damage:
+                        hero.health -= damage
+                    else:
+                        hero.delete()
+
+
+class DistanceEnemy(BaseEnemy):
+    def __init__(self, position, health, power):
+        super().__init__(position, health, power)
+
+    def action(self, hero):
+        if self.is_alive:
+            x1, y1, = self.rect.center
+            x2, y2 = hero.get_position()
+            vector = (x2 - x1, y2 - y1)
+            length = math.sqrt(vector[0] ** 2 + vector[1] ** 2)
+            shells.add(
+                Shell(self.rect.center, vector, length, False, self.power))
 
 
 class Game:  # служебный класс игры
@@ -326,6 +364,8 @@ class Game:  # служебный класс игры
         pygame.draw.rect(screen, pygame.Color("#464646"), frame_rect, width=2)
         print_text(f'{self.hero.health}/{MAX_HEALTH}', 70, 11, font_size=25)
 
+        print_text(f'SCORE: {current_score}', HALF_WIDTH + (HALF_WIDTH - get_message_size(f'SCORE: {current_score}', main_font, 35)[0]) - 10, 10, font_size=35)
+
         screen.blit(interface_images['armor'], (5, 45))
         frame_rect.y += 40
         points_rect.y += 40
@@ -342,13 +382,16 @@ class Game:  # служебный класс игры
         pygame.draw.rect(screen, pygame.Color("#464646"), frame_rect, width=2)
         print_text(f'{self.hero.mana}/{MAX_MANA}', 70, 91, font_size=25)
 
-        print_text(f'SCORE: {current_score}', HALF_WIDTH + (HALF_WIDTH - get_message_size(f'SCORE: {current_score}', main_font, 35)[0]) - 10, 10, font_size=35)
+        self.move_hero()
+        self.move_enemies()
+        for shell in shells:
+            self.move_shell(shell)
 
     def move_hero(self):
         keys = pygame.key.get_pressed()
 
         if any(keys):
-            x, y = self.hero.get_position()
+            x, y = self.hero.rect.topleft
             dx, dy = 0, 0
             if keys[pygame.K_w]:
                 dy -= self.hero.speed
@@ -358,34 +401,27 @@ class Game:  # служебный класс игры
                 if not dy:
                     dx -= self.hero.speed
                 else:
-                    dx -= self.hero.speed // 1.41
-                    dy //= 1.41
+                    dx -= self.hero.speed // 1.4
+                    dy //= 1.4
             if keys[pygame.K_d]:
                 if not dy:
                     dx += self.hero.speed
                 else:
-                    dx += self.hero.speed // 1.41
-                    dy //= 1.41
-
+                    dx += self.hero.speed // 1.4
+                    dy //= 1.4
             x += dx
             y += dy
 
             # герой не сможет зайти за границы borders
             if not pygame.sprite.spritecollideany(BaseEntity((x, y)), borders):
-                self.hero.rect.center = x, y
-                self.camera.update(self.hero)
-                for room in self.map.map:
-                    room.apply_camera(self.camera)
-                for enemy in enemies:
-                    enemy.apply_camera(self.camera)
-                    enemy.go_to((x, y))
-                for shell in shells:
-                    self.camera.apply(shell)
+                # self.hero.rect.center = x, y
+                self.camera.update(dx, dy)
+                self.apply_camera()
 
             # когда герой входит в какую-нибудь из комнат, комната
             # закрывается, пока её не зачистят
             if not self.in_room and pygame.sprite.spritecollideany(self.hero, entries):
-                entries.remove(*entries)
+                entries.empty()
                 self.in_room = True
                 self.lock_doors()
 
@@ -402,10 +438,18 @@ class Game:  # служебный класс игры
                 if not self.in_room:
                     self.change_doors_images('c', self.map.map[-1].entry, doors)
 
+    def apply_camera(self):
+        for room in self.map.map:
+            room.apply_camera(self.camera)
+        for enemy in enemies:
+            enemy.apply_camera(self.camera)
+        for shell in shells:
+            self.camera.apply(shell)
+
     def lock_doors(self):
         door_close.play()
         change_doors(doors)
-        doors.remove(*doors)
+        doors.empty()
         self.change_doors_images('c', self.map.map[-1].entry, current_doors)
         self.map.update()
         borders.add(*current_doors)
@@ -434,19 +478,47 @@ class Game:  # служебный класс игры
         borders.remove(*current_doors)
         self.change_doors_images('o', self.map.map[-3].entry, current_doors)
         for enemy in current_enemies:
-            enemy.delete()
-        current_doors.remove(*current_doors)
-        current_enemies.remove(*current_enemies)
+            enemy.kill()
+
+        current_doors.empty()
+        current_enemies.empty()
         self.in_room = False
         current_score += 1
 
     def move_enemies(self):
         for enemy in current_enemies:
-            enemy.move()
+            if isinstance(enemy, CloseEnemy):
+                enemy.move()
 
-    def move_shells(self):
-        for shell in shells:
-            shell.move()
+    def move_shell(self, shell):
+        x, y = shell.rect.center
+        x += shell.dx
+        y += shell.dy
+        if shell.from_hero:
+            for enemy in current_enemies:
+                if enemy.is_alive and pygame.sprite.collide_circle(shell, enemy):
+                    if enemy.health > shell.damage:
+                        enemy.health -= shell.damage
+                    else:
+                        enemy.delete()
+                    shells.remove(shell)
+        else:
+            if pygame.sprite.collide_circle(shell, self.hero):
+                if self.hero.armor > shell.damage:
+                    self.hero.armor -= shell.damage
+                else:
+                    damage = shell.damage - self.hero.armor
+                    self.hero.armor = 0
+                    if self.hero.health > damage:
+                        self.hero.health -= damage
+                    else:
+                        self.hero.delete()
+                shells.remove(shell)
+        if pygame.sprite.spritecollideany(shell, borders):
+            shells.remove(shell)
+
+        else:
+            shell.rect.center = x, y
 
 
 class Shell(pygame.sprite.Sprite):
@@ -461,25 +533,12 @@ class Shell(pygame.sprite.Sprite):
         self.rect = pygame.rect.Rect(x, y, 10, 10)
 
     def render(self, screen):
-        pygame.draw.circle(screen, pygame.Color("#2196f3"),
-                           self.rect.center, self.rect.width // 2)
-
-    def move(self):
-        x, y = self.rect.center
-        x += self.dx
-        y += self.dy
-        for enemy in current_enemies:
-            if pygame.sprite.collide_circle(self, enemy):
-                if enemy.health > self.damage:
-                    enemy.health -= self.damage
-                else:
-                    enemy.delete()
-                shells.remove(self)
-        if pygame.sprite.spritecollideany(self, borders):
-            shells.remove(self)
-
+        if self.from_hero:
+            pygame.draw.circle(screen, pygame.Color("#2196f3"),
+                               self.rect.center, self.rect.width // 2)
         else:
-            self.rect.center = x, y
+            pygame.draw.circle(screen, pygame.Color("Red"),
+                               self.rect.center, self.rect.width // 2)
 
 
 class Camera:
@@ -494,8 +553,5 @@ class Camera:
         obj.rect.y += self.dy
 
     # позиционировать камеру на объекте target
-    def update(self, target):
-        self.dx = -(target.rect.x + target.rect.w // 2 - WIDTH // 2)
-        self.dy = -(target.rect.y + target.rect.h // 2 - HEIGHT // 2)
-        target.rect.x += self.dx
-        target.rect.y += self.dy
+    def update(self, dx, dy):
+        self.dx, self.dy = - dx, - dy
